working DAC also integrated DAC into k2raster
