+++ raven/src/main/java/dk/itu/raven/ksquared/IntRank.java
@@ -126,9 +126,14 @@ public class IntRank implements Serializable {
             throw new IndexOutOfBoundsException("Invalid index: " + index);
         }
 
-        int blockIndex = index / BLOCK_SIZE;
-        int mask = vector[blockIndex] & (1 << (index % BLOCK_SIZE));
-        return mask != 0;
+        // int blockIndex = index >> 5;
+        // int mask = vector[blockIndex] & (1 << (index % BLOCK_SIZE));
+        // return mask != 0;
+        int windex = index >> 5;
+        int mask = 0x80000000 >>> (index & 0x1F);
+        int word = vector[windex];
+
+        return (word & mask) != 0;
     }
 
     public int size() {
