+++ raven/src/main/java/dk/itu/raven/util/LongArrayWrapper.java
@@ -2,9 +2,16 @@ package dk.itu.raven.util;
 
 public class LongArrayWrapper extends PrimitiveArrayWrapper {
     private long[] data;
+
     public LongArrayWrapper(long[] data) {
         this.data = data;
     }
+
+    public LongArrayWrapper(GoodLongArrayList data) {
+        this.data = new long[data.size];
+        System.arraycopy(data.array, 0, this.data, 0, data.size);
+    }
+
     public long get(int index) {
         return data[index];
     }
